reorganize openapi normalizer tests
